requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
multidict==6.0.3
py-cord==2.3.2
pydantic==1.10.2
python-decouple==3.6
SQLAlchemy==1.4.45
typing_extensions==4.4.0
yarl==1.8.2
